Move parseContent and prepareContent methods from Subs to Helpers
