+++ Sources/LightPortal/Helpers.php
@@ -590,6 +590,75 @@ class Helpers
 	}
 
 	/**
+	 * Prepare content to display
+	 *
+	 * Готовим контент к отображению в браузере
+	 *
+	 * @param string $content
+	 * @param string $type
+	 * @param int $block_id
+	 * @param int $cache_time
+	 * @return void
+	 */
+	public static function prepareContent(string &$content, string $type = 'bbc', int $block_id = 0, int $cache_time = 0)
+	{
+		global $context;
+
+		!empty($block_id) && !empty($context['lp_active_blocks'][$block_id])
+			? $parameters = $context['lp_active_blocks'][$block_id]['parameters'] ?? []
+			: $parameters = $context['lp_block']['options']['parameters'] ?? [];
+
+		Subs::runAddons('prepareContent', array(&$content, $type, $block_id, $cache_time, $parameters));
+	}
+
+	/**
+	 * Parse content depending on the type
+	 *
+	 * Парсим контент в зависимости от типа
+	 *
+	 * @param string $content
+	 * @param string $type
+	 * @return void
+	 */
+	public static function parseContent(string &$content, string $type = 'bbc')
+	{
+		global $context;
+
+		switch ($type) {
+			case 'bbc':
+				$content = parse_bbc($content);
+
+				// Integrate with the Paragrapher mod
+				call_integration_hook('integrate_paragrapher_string', array(&$content));
+
+				break;
+			case 'html':
+				$content = un_htmlspecialchars($content);
+
+				break;
+			case 'php':
+				$content = trim(un_htmlspecialchars($content));
+				$content = trim($content, '<?php');
+				$content = trim($content, '?>');
+
+				ob_start();
+
+				try {
+					$content = html_entity_decode($content, ENT_COMPAT, $context['character_set'] ?? 'UTF-8');
+					eval($content);
+				} catch (\ParseError $p) {
+					echo $p->getMessage();
+				}
+
+				$content = ob_get_clean();
+
+				break;
+			default:
+				Subs::runAddons('parseContent', array(&$content, $type));
+		}
+	}
+
+	/**
 	 * Get the filtered $obj[$key]
 	 *
 	 * Получаем отфильтрованное значение $obj[$key]
